backend/requirement.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-seed==0.3.1
djangorestframework==3.15.1
Faker==25.4.0
mysqlclient==2.2.4
pillow==10.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
toposort==1.10
tzdata==2024.1
